% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item_data.R
\name{create_item_data}
\alias{create_item_data}
\title{Create a tibble of information about the object}
\usage{
create_item_data(.x)
}
\arguments{
\item{.x}{An object returned from \code{\link[=osrs_api]{osrs_api()}}}
}
\value{
A tibble with information about items, prices, dates
}
\description{
Create a tibble of information about the object
}
\examples{
# Get the entire price history of the abyssal whip
whip <- osrs_api(items = "Abyssal_whip", history = "all")

# Create a data set with information about the item
create_item_data(whip)

# Same can be done for other values of history
osrs_api(items = c("bucket", "Abyssal whip"), history = "latest") |>
  create_item_data()
osrs_api(items = c("Dragon dagger"), history = "sample") |>
  create_item_data()

}
