% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{osrs_api}
\alias{osrs_api}
\title{Get information about Grand Exchange prices of items in OSRS}
\usage{
osrs_api(item, history = "latest", user_agent = "GE_price_tracker")
}
\arguments{
\item{item}{\code{character} name of the item}

\item{history}{`character` [mean()] with type of request. Default is "latest", but "all" is
also available}

\item{user_agent}{`character` of the user agent}
}
\value{
A list of class [osrs_api]
}
\description{
Use the Weird Gloop API to fetch information about Grand Exchange prices in
Oldschool Runescape
}
\examples{
osrs_api(item = "Abyssal_whip", history = "latest")
}
