% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrsAPI-package.R
\docType{package}
\name{osrsAPI-package}
\alias{osrsAPI}
\alias{osrsAPI-package}
\title{osrsAPI: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mathiaslj.github.io/osrsAPI/}
}

}
\keyword{internal}
