% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.osrs_api}
\alias{print.osrs_api}
\title{Print method for class}
\usage{
\method{print}{osrs_api}(x, ...)
}
\arguments{
\item{x}{object of class `osrs_api`}

\item{...}{additional arguments passed to methods}
}
\description{
Print method for class
}
\examples{
whip <- osrs_api(item = "Abyssal_whip", history = "latest")
print(whip)
}
