% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{osrs_api}
\alias{osrs_api}
\title{Get information about Grand Exchange prices of items in OSRS}
\usage{
osrs_api(items, history = "latest", user_agent = "GE_price_tracker")
}
\arguments{
\item{items}{\code{character} vector of the items}

\item{history}{\code{character} with type of request. Default is "latest", but "all" is
also available}

\item{user_agent}{\code{character} of the user agent}
}
\value{
A list of class \code{osrs_api}
}
\description{
Use the Weird Gloop API to fetch information about Grand Exchange prices in
Oldschool Runescape
}
\examples{
# Get the latest prices of both a bucket and the abyssal whip
bucket_and_whip <- osrs_api(items = c("bucket", "Abyssal whip"), history = "latest")

# Get the entire price history of the abyssal whip
whip <- osrs_api(items = "Abyssal_whip", history = "all")

}
